Add installer concept.
