+++ source/view/installer/setup.cpp
@@ -1,0 +1,6 @@
+#include "setup.hpp"
+
+Setup::Setup()
+{
+
+}
