+++ source/view/installer/setup.hpp
@@ -1,0 +1,10 @@
+#ifndef SETUP_VIEW_HPP
+#define SETUP_VIEW_HPP
+
+class Setup
+{
+public:
+    Setup();
+};
+
+#endif // SETUP_VIEW_HPP
